% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.R
\name{ddm}
\alias{ddm}
\title{Fitting function for the DDM}
\usage{
ddm(
  drift,
  boundary = ~1,
  ndt = ~1,
  bias = 0.5,
  sv = 0,
  data,
  args_method = list(),
  args_ddm = list(err_tol = 1e-06, switch_thresh = 0.8),
  use_gradient = TRUE
)
}
\arguments{
\item{drift}{Formula for the drift rate (v)}

\item{boundary}{Formula for the boundary separation (a).}

\item{bias}{Formula for the relative initial bias (w).}

\item{sv}{Formula for the inter-trial variability in the drift rate (sv).}

\item{ntd}{Formula for the non-decision time (t0).}
}
\value{
Currently, it's just the output from nlminb.
}
\description{
This function fits the DDM to the data that is provided, using formulas for
the DDM parameters if desired.
}
\examples{
# prepare data
data(med_dec, package = "fddm")
med_dec <- med_dec[which(med_dec[["rt"]] >= 0), ]
p1 <- med_dec[med_dec[["id"]] == 2 &
                      med_dec[["group"]] == "experienced", ]
p1[["response"]] <- as.factor(p1[["response"]])

# example fit with gradient information
ddm(rt + response ~ 0 + classification:difficulty, data = p1)

# example fit without gradient without gradient information
ddm(rt + response ~ 0 + classification:difficulty, data = p1,
    use_gradient = FALSE)
}
