% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_dec-data.R
\docType{data}
\name{med_dec}
\alias{med_dec}
\title{Medicial decision data}
\format{A data frame with 11000 rows and 9 variables:
\describe{
  \item{id}{identification number of the participant}
  \item{group}{expertise of participant; either "expert" or "novice"}
  \item{block}{block number}
  \item{trial}{index of trial for each participant}
  \item{classification}{true classification of the pictured cell; i.e. the correct response}
  \item{difficulty}{adjudged difficulty of the task for the particular image}
  \item{response}{response given by the participant; either "blast" or "non-blast"}
  \item{rt}{the response time associated with the response, in seconds}
  \item{stimulus}{the image file used for the specific trial}
}}
\source{
Trueblood, J.S., Holmes, W.R., Seegmiller, A.C. et al. The impact of
  speed and bias on the cognitive processes of experts and novices in medical
  image decision-making. Cogn. Research 3, 28 (2018).
  https://doi.org/10.1186/s41235-018-0119-2
}
\usage{
med_dec
}
\description{
A dataset containing the binary responses and associated response times for
an experiment where experts and novices were presented with images of cells
and each participant had to decide whether the pictured cell was "blast" or
"non-blast"
}
\keyword{dataset}
