% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfddm.R
\name{dfddm}
\alias{dfddm}
\title{Density of Ratcliff Diffusion Decision Model}
\usage{
dfddm(
  rt,
  response,
  a,
  v,
  t0,
  w = 0.5,
  sv = 0,
  log = FALSE,
  n_terms_small = c("Foster", "Navarro", "Kesselmeier"),
  summation_small = c("2017", "2014"),
  scale = c("small", "large", "both"),
  eps = 1e-06
)
}
\arguments{
\item{rt}{a vector of RTs.}

\item{response}{???}

\item{a}{threshold separation. Amount of information that is considered for a
decision. Large values indicate a conservative decisional style. Typical
range: 0.5 < \code{a} < 2}

\item{v}{drift rate. Average slope of the information accumulation process.
The drift gives information about the speed and direction of the
accumulation of information. Large (absolute) values of drift indicate a
good performance. If received information supports the response linked to
the upper threshold the sign will be positive and vice versa. Typical
range: -5 < \code{v} < 5}

\item{t0}{non-decision time or response time constant (in seconds). Lower
bound for the duration of all non-decisional processes (encoding and
response execution). Typical range: 0.1 < \code{t0} < 0.5}

\item{w}{relative starting point. Indicator of an a priori bias in decision
making. When the relative starting point \code{w} deviates from
\code{0.5}, the amount of information necessary for a decision differs
between response alternatives. Default is \code{0.5} (i.e., no bias).}

\item{sv}{inter-trial-variability of drift rate. Standard deviation of a
normal distribution with mean \code{v} describing the distribution of
actual drift rates from specific trials. Values different from 0 can
predict slow errors. Typical range: 0 < \code{sv} < 2. Default is 0.}

\item{log}{logical; if \code{TRUE}, probabilities p are given as log(p).}

\item{n_terms_small}{Method for calculating number of terms used in the sum
of the small time approximation of the DDM.}

\item{summation_small}{Method}

\item{scale}{some stuff}

\item{eps}{\code{numerical} scalar value. Precision of calculation.}
}
\value{
Density of the DDM with precision \code{eps} of length \code{rt}.
}
\description{
Density function for the Ratcliff diffusion decision model (DDM) with
following parameters: \code{a} (threshold separation), \code{w} (relative
starting point), \code{v} (drift rate), \code{t0} (non-decision time/response
time constant), and \code{sv} (inter-trial-variability of drift).
}
\details{
The different methods implemented
}
\examples{

dfddm(c(1, 2, 3), 1, a = 1, v = 2, t0 = 0.3)
}
\references{
all the papers go here
}
